Create Bubble Sort Algorithm
This algorithm works by repeatedly comparing adjacent elements in the aarray and swapping them if they are in the wrong order. The algorithm starts at the beginning of the arrray and compares the first two elements. If the first element is greater then the second element, they are swapped.

The bubble sort algorithm is a simple sorting algorithm, but it is not very efficient. It is O(n^2) in time complexity., which means that the time is takes to sort an array increases quadratically with the size of the array.